requirements.txt
kafka-python
pymongo
psycopg2
pandas
requests
streamlit
uvicorn
fastapi
sqlalchemy